mspurchasing/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==3.1.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
docopt==0.6.2
entrypoints==0.4
executing==1.0.0
filelock==3.12.0
idna==3.4
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==3.5.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==23.2.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.6.3
urllib3==2.0.2
virtualenv==20.23.0
wcwidth==0.2.5
yarg==0.1.9
